### 2018-10-02 #255 @stijnvanhoey
wateRinfo package
wateRinfo facilitates the access to a variety of environmental water-related data about Flanders (Belgium) available on [waterinfo.be](https://www.waterinfo.be/), a website managed by the Flemish Environmental Agency.

```
Package: wateRinfo
Title: Download Time Series Data from Waterinfo.be
Version: 0.2.0
Description: The wateRinfo package provides an R interface to the data available
    at waterinfo.be, the data portal provided by the Flemish Environmental
    Agency. The package provides R functions to check the available stations for
    a given variable, check the available variable for a given station and
    download time series using the database identifiers as well as station
    names.
    person("Stijn", "Van Hoey", role = c("aut", "cre"), email = "stijn.vanhoey@inbo.be", comment = c(ORCID = "0000-0001-6413-3185")),
    person("Willem", "Maetens", role = "ctb", email = "w.maetens@vmm.be"),
    person("Peter", "Desmet", role = "ctb", email = "peter.desmet@inbo.be", comment = c(ORCID = "0000-0002-8442-8025"))
    )
    httr,
    openssl,
    lubridate (>= 1.6.0),
    utils
    covr,
RoxygenNote: 6.1.0.9000
```

https://github.com/inbo/wateRinfo

**data retrieval**: because the package provides R access to environmental data (water level and tidal information, discharge data, water quality variables, meteorological variables...) provided by the Flemish Environmental Agency. Without the package, people need to download data manually using the web interface or using the available API calls. The web interface, although evolving, does not support downloads for a list of stations and variables leading to lots of clicks to download data. The [API documentation ](https://www.waterinfo.be/download/9f5ee0c9-dafa-46de-958b-7cac46eb8c23?dl=0)is limited and only available in Dutch. The package overcomes these limitations, supporting downloads of multiple variables and stations more efficiently.

Anyone interested in using the water-related data from waterinfo.be, for example hydrologists and ecologists. Scientific applications are ranging from environmental prediction and risk assessment modelling studies (e.g hydrological flood models) to ecological studies requiring environmental data.

To our knowledge, there is no other R package to retrieve data from waterinfo.be. Although some loose scripts have been circulating among researchers from different institutes, this package aims to support a collaborative/community effort.

    * NEWS file - there is one now, but only started with version 0.1.1 and using the ropensci structure since 0.2.0
    * Package name: Unfortunately, the name contains capital letters (wateRinfo), but this is how the package is already in use for almost a year
    * The main download function names are setup as `get_**` which does not use the `object_verb()` recommendation. As these are the main functions, we would keep the naming as such.
    * We did not use the `@family` tag as this provides full control on the order in between different sections. Considering the limited set of functionalities, the management of the `pkgdown` yml file is still feasible.
    * We do run the documentation website locally before deployment.
    * The current set of untested functionalities consists of interactions (mainly exceptions) with the waterinfo.be API which are hard to replicate.
