### 2018-10-03 #255 @sckott
@stijnvanhoey et al. Do other regions of Belgium have water data available?  If so, can the pkg be generalized to use those other data?
