### 2018-10-03 #255 @stijnvanhoey
Good point. I'll try to provide an overview of the information I'm aware of. Notice that we have for water-related issues different governmental levels: national level and regional levels, i.e. Flanders, Walloon and Brussels). Rivers are monitored on regional (and subregional) level:
- The main data source for meteorological data (rainfall, evapotranspiration,...) on the national level is the [Royal Meteorological Institute of Belgium](http://www.meteo.be/meteo/view/en/65239-Home.html), but they do not provide open data for the moment.
- For the Walloon region, there is no data on the [open data platform](https://opendata.digitalwallonia.be/dataset?q=&tags=Environnement&sort=score+desc%2C+metadata_modified+desc), but there data on discharge on their [aqualim](http://aqualim.environnement.wallonie.be/login.do?time=1538594260100) website. The latter presents data from the [last 30 days online](http://aqualim.environnement.wallonie.be/Station.do?method=selectStation&time=1538594305999&station=L5201) and a form to [download data as an excel-sheet](http://aqualim.environnement.wallonie.be/Station.do?method=displayDownloadData&sParam=HAUT&typeGraph=GRAPH_HORAIRE&oper=H30&time=1538594384138) when providing contact details etc. For water quality, there are two websites: [aquapol](http://aquapol.environnement.wallonie.be/StationDataHisto.do?method=tableData) and [aquaphyc](http://aquaphyc.environnement.wallonie.be/GeneralPages.do?method=displayAccueil&menu1=accueil) providing water quality data for a number of stations. I do not have a direct idea how these could be incorporated (no REST or something alike services)?
- For Brussels, the is no water data available on the [open data platform](https://opendata.brussel.be/page/home/). Their main river is *Zenne*, see the [map](https://leefmilieu.brussels/themas/water/water-brussel/het-blauwe-netwerk/kaart-van-de-waterlopen). I do know of a project [belini](http://life-belini.be/portfolio-item/monitoring/) with a monitoring component, but the data seems not available as open data or a webservice.
- The Flemish environmental agency (VMM) also provides data about [water quality](http://geoloket.vmm.be/Geoviews/), but they told us their water quality system is currently in redevelopment (with other endpoints), so rather something to take into scope when the new system will be in place.
- There is a database/webportal on soil and groundwater in Flanders, called [DOV](https://www.dov.vlaanderen.be/), which is, among other subsoil data (e.g. geological data), the aggregator for groundwater time series. There is a current initiative to increase the number of data contributors to DOV and a Python package for data access to DOV is currently in development as well, https://github.com/dov-vlaanderen/pydov. As the groundwater level data is thematically (and technically) linked to the other data sources of DOV (geological interpretations, boreholes,...) we consider it out of scope for the wateRinfo package.

Notice that we opt for GBIF to publish species occurrences related to water, e.g. [fish](https://www.gbif.org/dataset/823dc56e-f987-495c-98bf-43318719e30f), so this data is covered by the rgbif package.

As such, to my knowledge, there is no direct other source of data to incorporate. I'm adding @WillemMaetens (VMM), @pietercolpaert (open knowledge), @Sachagobeyn, @bartpannemans, @pjhaest to the discussion,  maybe they know about other initiatives?
