### 2018-10-04 #255 @Sachagobeyn
Dear all,

With respect to water quality data hosted by the VMM (Flanders): as I am currently aware, they are not available open source. They can be requested via the VMM website, however from my own experience, requests are often declined. An alternative for this type of information is the European Union Open Data Portal, in which ecological water quality, and a number of physico-chemical water quality variables are reported. For the latter, it should noted that data are only available for a few years. In addition, the datasets are not 'callable' as they are in waterinfo (zips hosted on the website).

http://cdr.eionet.europa.eu/be
https://www.eea.europa.eu/data-and-maps
