+++ .travis.yml
@@ -1,4 +1,4 @@
-# R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
+# R for travis: see documentation at Travis docs https://docs.travis-ci.com/user/languages/r
 
 language: r
 sudo: false
