### 2018-10-17 07fa3ec
Update vignette information on datasource
