+++ vignettes/download_timeseries.Rmd
@@ -92,7 +92,7 @@ pet_yearly %>%
 
 (see the vignette on defining date periods for more information on the period argument used)
 
-**Remark:** the `get_stations` function only works for those measurement stations belonging to the VMM `meetnet` (network), related to the so-called `datasource = 1`. For [other networks](http://www.waterinfo.be/default.aspx?path=NL/HIC/Recent_toelichting), i.e. `datasource = 2`, the enlisting is not supported. Still, a search for data is provided starting from a given station name, as explained in the next section.
+**Remark:** the `get_stations` function only works for those measurement stations belonging to the VMM `meetnet` (network), related to the so-called `datasource = 1`. For [other networks](http://www.waterinfo.be/default.aspx?path=NL/HIC/Recent_toelichting), i.e. `datasource = 4`, the enlisting is not supported. Still, a search for data is provided starting from a given station name, as explained in the next section.
 
 ## Search identifier based on station name
 
@@ -108,14 +108,14 @@ available_variables %>% select(ts_id, station_name, ts_name,
                                parametertype_name)
 ```
 
-The available number of variables depends on the measurement station. The representation is not standardized and depends also from the type of [`meetnet`](http://www.waterinfo.be/default.aspx?path=NL/HIC/Recent_toelichting). Nevertheless, one can derive the required `ts_id` from the list when interpreting the field names. Remark that the datasource can be 2 instead of 1 for specific `meetnetten` (networks). The datasource to use is printed when asking the variables for a station. 
+The available number of variables depends on the measurement station. The representation is not standardized and depends also from the type of [`meetnet`](http://www.waterinfo.be/default.aspx?path=NL/HIC/Recent_toelichting). Nevertheless, one can derive the required `ts_id` from the list when interpreting the field names. Remark that the datasource can be 4 instead of 1 for specific `meetnetten` (networks). The datasource to use is printed when asking the variables for a station. 
 
-In order to download the 10 min time series water level data for the station in Sint-Amands tij/Zeeschelde, the `ts_id = 2810011` can be used in the `get_timeseries_tsid` function, taking into account the `datasource = 2` (default is 1):
+In order to download the 10 min time series water level data for the station in Sint-Amands tij/Zeeschelde, the `ts_id = 2810011` can be used in the `get_timeseries_tsid` function, taking into account the `datasource = 4` (default is 1):
 
 ```{r downloadstamands, fig.width = 7}
 tide_stamands <- get_timeseries_tsid("2810011", 
                                      from = "2017-06-01", to = "2017-06-05",
-                                     datasource = 2)
+                                     datasource = 4)
 ggplot(tide_stamands, aes(Timestamp, Value)) + 
     geom_line() + xlab("") + ylab("waterlevel")
 ```
