### 2018-10-17 e5ec1b7
Update tests to fit new identifiers
