### 2018-11-01 #255 @karthik
### Editor checks:

---

#### Editor comments

👋 @stijnvanhoey. Apologies for the delay in assigning a reviewer (just returning from travels). Laura has kindly agreed to review. Given that we have a week off in November, I have given her a month to review this submission. I have also posted some results from the goodpractice check (from the `goodpractice` package). Please address these issues as they will also come up in the review.

```
---
── GP wateRinfo ────────────────────────────────────────────────────────────────

It is good practice to

    R/token.R:44:1
    tests/testthat/test-token_usage.R:75:1
    tests/testthat/test-waterinfo_call.R:35:1

  ✖ fix this R CMD check ERROR: Running examples in
    ‘wateRinfo-Ex.R’ failed The error most likely occurred in: > ###
    Name: get_timeseries_tsid > ### Title: Download timeseries data
    from waterinfo.be > ### Aliases: get_timeseries_tsid > > ### **
    Examples > > get_timeseries_tsid("35055042", from = "2017-01-01",
    to = "2017-01-02") Timestamp Value Quality Code 1 2016-12-31
    23:00:00 0.00 130 2 2016-12-31 23:15:00 0.00 130 3 2016-12-31
    23:30:00 0.00 130 4 2016-12-31 23:45:00 0.00 130 5 2017-01-01
    00:00:00 0.00 130 6 2017-01-01 00:15:00 0.00 130 7
    ...
    Error: Waterinfo API request failed [500] Waterinfo error:
    InvalidParameterValue Waterinfo return message: Datasource
    parameter not found in config. Execution halted
  ✖ checking tests ... ERROR Running the tests in
    ‘tests/testthat.R’ failed. Last 13 lines of output: >
    library(testthat) > library(wateRinfo) > > test_check("wateRinfo")
    ── 1. Failure: non existing tsid to API (@test-waterinfo_call.R#21)
    ─────────── `call_waterinfo(query)` threw an error with unexpected
    message. Expected match: "Waterinfo API request
    failed.*InvalidParameterValue" Actual message: "API did not return
    json - " ══ testthat results
    ═══════════════════════════════════════════════════════════ OK: 104
    SKIPPED: 4 FAILED: 1 1. Failure: non existing tsid to API
    (@test-waterinfo_call.R#21) Error: testthat unit tests failed
    Execution halted
  ✖ fix this R CMD check WARNING: LaTeX errors when creating
    PDF version. This typically indicates Rd problems.
  ✖ fix this R CMD check ERROR: Re-running with no
    redirection of stdout/stderr. Hmm ... looks like a package You may
    want to clean up by 'rm -Rf /tmp/RtmpHVX5D6/Rd2pdf9fb1c6e9df4'
─────────────────────────────────────────────────────────────────────────
```

---

Reviewers: @ldecicco-usgs
Due date: Nov 29, 2018
