### 2018-11-07 #255 @stijnvanhoey
Thanks for the information.

With respect to the issues with the example and unit tests: Apparently, the unique time series identifiers have changed on the [waterinfo.be](https://www.waterinfo.be/) side for a subset of stations. I was expecting them to be stable, but they are not. I'm doing the best I can to fix these issues asap.
