### 2018-11-07 456daec
Add proper error handling on tomcat server isues
