### 2018-11-07 b87036a
Split long code lines over multiple lines
