### 2018-11-07 #255 @stijnvanhoey
New version `fix-source-info` branch (will be merged soon), has following output for the `gp` command:

```
── GP wateRinfo ─────────────────────────────────────────────────────────────────────────────────────────────────────────

It is good practice to

  ✖ write unit tests for all functions, and all package code in general. 87% of code lines are
    covered by test cases.

    R/call_waterinfo.R:58:NA
    R/call_waterinfo.R:59:NA
    R/call_waterinfo.R:60:NA
    R/call_waterinfo.R:61:NA
    R/call_waterinfo.R:62:NA
    ... and 49 more lines

  ✖ fix this R CMD check WARNING: LaTeX errors when creating PDF version. This typically
    indicates Rd problems.
─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
```

One of the elements missing in the unit testing is the XML-parsing of a specific type of error. We encountered the issue earlier, but it is hard to reproduce systematically. Hence, we do not know exactly how to properly handle this. Advice is welcome.
