### 2018-11-07 #255 @stijnvanhoey
With respect to the Tex errors, I'm not completely aware on how to tackle this. I tried to execute  `devtools check` with the manual creation option:
```
devtools::check(document = FALSE, cleanup = TRUE, manual = TRUE)
```
I get the warning:

```
❯ checking PDF version of manual ... WARNING
  LaTeX errors when creating PDF version.
  This typically indicates Rd problems.
```

However, when I do check the logs and output in the `wateRinfo.Rcheck` folder, the `wateRinfo-manual.pdf` is there and the `Rdlatex.log` is (I just subset some potentially relevant sections):

```
Hmm ... looks like a package
This is pdfTeX, Version 3.14159265-2.6-1.40.17 (TeX Live 2016/Debian) (preloaded format=pdflatex)
 restricted \write18 enabled.
entering extended mode
(./Rd2.tex
LaTeX2e <2016/03/31>
Babel <3.9r> and hyphenation patterns for 5 language(s) loaded.

...
No file Rd2.aux.
...
No file Rd2.toc.
...
Overfull \hbox (15.18042pt too wide) in paragraph at lines 370--370
...

Output written on Rd2.pdf (12 pages, 90183 bytes).
Transcript written on Rd2.log.
Saving output to ‘wateRinfo-manual.pdf’ ...
Done
You may want to clean up by 'rm -rf /tmp/RtmpeypCkA/Rd2pdf508051cc1987'
```
And the output looks like: [wateRinfo-manual.pdf](https://github.com/ropensci/onboarding/files/2557573/wateRinfo-manual.pdf) Another file, `wateRinfo-Ex.pdf` is also created, but *document contains no pages*.

Any input on how to tackle this issue is certainly welcome.
