### 2018-11-07 2cb0cf0
Update links in roxygen docs
