### 2018-11-13 e9c4fe1
Build site for version 0.2.2 (with new logo)
