### 2018-11-28 32b47f6
Add "INBO" acronym in full name to be consistent
