+++ DESCRIPTION
@@ -11,7 +11,7 @@ Authors@R: c(
     person("Stijn", "Van Hoey", role = c("aut", "cre"), email = "stijn.vanhoey@inbo.be", comment = c(ORCID = "0000-0001-6413-3185")),
     person("Willem", "Maetens", role = "ctb", email = "w.maetens@vmm.be"),
     person("Peter", "Desmet", role = "ctb", email = "peter.desmet@inbo.be", comment = c(ORCID = "0000-0002-8442-8025")),
-    person("Research Institute for Nature and Forest", role = c("cph"), email = "info@inbo.be")
+    person("Research Institute for Nature and Forest (INBO)", role = c("cph"), email = "info@inbo.be")
     )
 License: MIT + file LICENSE
 URL: https://github.com/inbo/wateRinfo, https://inbo.github.io/wateRinfo
