### 2018-11-28 8fc0cc1
Update readme with token necessity informationgs
