### 2018-11-28 aefecbe
Provide char instead of factor by default
