+++ R/get_variables.R
@@ -63,10 +63,12 @@ get_variables <- function(station_no, token = NULL) {
 
   stations <- station_variables$content
   if (dim(stations)[1] == 2) {
-    df <- as.data.frame(t(stations[2:nrow(stations), ]))
+    df <- as.data.frame(t(stations[2:nrow(stations), ]),
+                        stringsAsFactors = FALSE)
     colnames(df) <- stations[1, ]
   } else {
-    df <- as.data.frame(stations[2:nrow(stations), ])
+    df <- as.data.frame(stations[2:nrow(stations), ],
+                        stringsAsFactors = FALSE)
     colnames(df) <- stations[1, ]
   }
 
