+++ R/get_stations.R
@@ -33,6 +33,8 @@
 #'   \item{ts_unitsymbol}{Unit of the variable.}
 #'   \item{dataprovider}{Data provider of the time series data.}
 #' }
+#' The URL of the specific request is provided as a comment attribute to the
+#' returned data.frame. Use \code{comment(df)} to get the request URL.
 #'
 #' @export
 #' @importFrom dplyr %>% select
