+++ R/get_variables.R
@@ -22,6 +22,8 @@
 #'   \item{parametertype_name}{Measured variable description.}
 #'   \item{stationparameter_name}{Station specific variable description.}
 #' }
+#' The URL of the specific request is provided as a comment attribute to the
+#' returned data.frame. Use \code{comment(df)} to get the request URL.
 #'
 #' @export
 #'
