### 2018-11-28 c6e04b5
Add unit tests for URL as comment attribute
