+++ tests/testthat/test-timeseries_download.R
@@ -9,6 +9,11 @@ test_that("Call works as expected", {
   )
 
   expect_is(pressure, "data.frame")
+
+  expect_is(comment(pressure), "character")
+  expect_true(grepl("http://download.waterinfo.be/tsmdownload/KiWIS/KiWIS",
+                    comment(pressure)))
+
   expect_equal(
     colnames(pressure),
     c("Timestamp", "Value", "Quality Code")
@@ -27,6 +32,8 @@ test_that("Response of empty dataframe", {
   )
 
   expect_is(pressure, "data.frame")
+  expect_is(comment(pressure), "character")
+
   expect_equal(
     colnames(pressure),
     c("Timestamp", "Value", "Quality Code")
