### 2018-11-28 65e0441
Add test for variable output char data types
