### 2018-11-28 6184839
Use vapply instead of sapply
