### 2018-11-29 1d215ef
Indicate to knit README.Rmd
