+++ README.Rmd
@@ -2,7 +2,7 @@
 output: github_document
 ---
 
-<!-- README.md is generated from README.Rmd. Please edit that file -->
+<!-- README.md is generated from README.Rmd. Please edit that file and knit -->
 
 ```{r, echo = FALSE}
 knitr::opts_chunk$set(
