### 2018-11-29 8569186
Add high res logo
