### 2018-11-29 6c94560
Close img tag correctly
