### 2018-12-06 7732f01
Adjust links to https links
