### 2018-12-06 #255 @karthik
Congrats @stijnvanhoey, your submission has been approved! 🎉 Thank you for submitting and @ldecicco-USGS for thorough and timely review. I'm very pleased with this package and find it very easy to use and well put together. To-dos:

```
[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)

```
- [ ] Fix any links in badges for CI and coverage to point to the ropensci URL. (We'll turn on the services on our end as needed)

Welcome aboard! We'd also love a blog post about your package, either a short-form intro to it (https://ropensci.org/technotes/) or long-form post with more narrative about its development. ((https://ropensci.org/blog/). If you are, @stefaniebutland will be in touch about content and timing.
