### 2018-12-10 #255 @stijnvanhoey
@karthik I transferred the repo, but apparently I should have added the `wateRinfo` team during transfer. Sorry to ask you again, could you provide the waterinfo team admin access to the repo? Thanks
