### 2018-12-10 #255 @stefaniebutland
@stijnvanhoey Your draft post will be great for the rOpenSci blog! Such a good weaving of practical and very engaging story with package use case. That video is incredible. We can publish on Jan 15, 2019.

Please submit your draft as a pull request by Jan 8 according to the instructions here: https://github.com/ropensci/roweb2#contributing-a-blog-post. That will give me an opportunity to review prior to publication.

Looking forward to this one!
