### 2018-12-11 dd3c129
Add ropensci footer
