### 2018-12-11 4c7428e
Transfer links to ropensci
