+++ README.Rmd
@@ -26,7 +26,7 @@ You can install wateRinfo from GitHub with:
 
 ```{r gh-installation, eval = FALSE}
 # install.packages("devtools")
-devtools::install_github("inbo/wateRinfo")
+devtools::install_github("ropensci/wateRinfo")
 ```
 
 When succesfull, load it as usual:
@@ -148,6 +148,6 @@ This package is an experimental open-source project. All functionalities are pro
 
 Please note that this project is released with a [Contributor Code of Conduct](.github/CODE_OF_CONDUCT.md). By participating in this project you agree to abide by its terms.
 
-Please [report any issues or bugs](https://github.com/inbo/wateRinfo/issues/new).
+Please [report any issues or bugs](https://github.com/ropensci/wateRinfo/issues/new).
 
 [![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
