### 2018-12-11 #255 @stijnvanhoey
@stefaniebutland Great! We'll make sure to have a version submitted by Jan 8.
@karthik thanks for all the help, a PR is ready, just a single check on the coveralls link, https://github.com/ropensci/wateRinfo/pull/47

Any other admin we need to do or can I/you close this issue?
