### 2018-12-11 e2d7dfe
Update some more URLs to ropensci
