+++ _pkgdown.yml
@@ -19,7 +19,7 @@ navbar:
   - icon: fa-twitter fa-lg
     href: https://twitter.com/LifeWatchINBO
   - icon: fa-github fa-lg
-    href: https://github.com/inbo/wateRinfo
+    href: https://github.com/ropensci/wateRinfo
 
 home:
   links:
