+++ _pkgdown.yml
@@ -31,6 +31,8 @@ home:
 authors:
   Stijn Van Hoey:
     href: https://twitter.com/SVanHoey
+  Peter Desmet:
+    href: https://twitter.com/peterdesmet
 
 reference:
   - title: "Main download functionalities"
