### 2018-12-13 c0df034
Bump to 0.3.0 with dev notice
