### 2019-01-08 #255 @stijnvanhoey
@stefaniebutland, we are running into deadline trouble with the blog post. The current version is available [here](https://github.com/stijnvanhoey/wateRinfo_post), but we still need to have some internal rewriting and review.  Would it still be ok to submit the PR on Monday? Our excuses for the rather late notice.
