### 2019-01-08 #255 @stefaniebutland
Not a problem @stijnvanhoey. We're very happy to host your post when it is ready.

If you submit a pull request on Monday I'll do my best to do a final review asap after it's in. My responses might be delayed, however, because I'm involved in a workshop next week. The post is in very good shape so I don't expect it will need many changes after my review.

If you're not able to submit on Monday, then submit when ready and we can assign a different publication date.

rOpenSci will tweet about the post and you are welcome to suggest any tweet-content, preferred images or people who should be tagged and I will work to incorporate them.

Thank you for getting in touch!
